+++ routes/users.ts
@@ -1,0 +1,31 @@
+import { NextFunction, Request, Response } from 'express';
+import express from 'express';
+const router = express.Router();
+
+import User from '../db/schema/user';
+
+type FIX_ME = unknown;
+
+router.get('/users', (req: Request, res: Response) => {
+  User.find({}).then(function (users: FIX_ME) {
+    res.send(users);
+  });
+});
+
+router.post('/users', (req: Request, res: Response, next: NextFunction) => {
+  const newUser = {
+    uuid: req.body.uuid,
+    userName: req.body.userName,
+    password: req.body.password,
+    preferredPaymentMethod: req.body.preferredPaymentMethod,
+  };
+  User.create(newUser)
+    .then(function (user: FIX_ME) {
+      res.send(user);
+    })
+    .catch((e: unknown) => {
+      next(e);
+    });
+});
+
+export const users = router;
